Move converter state management to session
